added caching, reorganized some mongo stuff
