+++ server/package.json
@@ -18,11 +18,12 @@
     "mongoose": "^5.11.15",
     "node": "^15.7.0",
     "nodemailer": "^6.4.17",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.3"
   },
   "devDependencies": {
     "@types/node": "^14.14.25",
+    "@types/mongoose": "^5.10.3",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3",
     "faker": "^5.3.1",
     "nodemon": "^2.0.7"
   }
